+++ cogs/info.py
@@ -2,8 +2,9 @@ import discord
 from discord import app_commands
 from discord.ext import commands
 from db_actions import *
-from buttons import *
 import random
+import datetime
+from views.help import *
 
 # Commands which give general info
 class InfoCog(commands.Cog):
