+++ views/confirmprompt.py
@@ -1,0 +1,25 @@
+import discord
+
+class ConfirmPrompt(discord.ui.View):
+    def __init__(self, user):
+        super().__init__(timeout=60)
+        self.confirmed = None
+        self.user = user
+
+    async def interaction_check(self, interaction: discord.Interaction):
+        if interaction.user != self.user:
+            await interaction.response.defer()
+            return False
+        return True
+
+    @discord.ui.button(label="Confirm", style=discord.ButtonStyle.green)
+    async def confirm(self, interaction: discord.Interaction, button: discord.ui.Button):
+        self.confirmed = True
+        await interaction.response.defer()
+        self.stop()
+    
+    @discord.ui.button(label="Cancel", style=discord.ButtonStyle.red)
+    async def deny(self, interaction: discord.Interaction, button: discord.ui.Button):
+        self.confirmed = False
+        await interaction.response.defer()
+        self.stop()
